requirements.txt
numpy
scipy
faiss-cpu
scikit-learn
requests
jsonschema
groq
tqdm
joblib
langchain
langchain-core
langchain-community
langchain-groq
pydantic
python-dotenv